refactor model classes into separate files
