+++ base_agent/ttad/ttad_transformer_model/train_model.py
@@ -19,6 +19,10 @@ from transformers import AutoModel, AutoTokenizer, BertConfig
 
 from utils_parsing import *
 from utils_caip import *
+from decoder_with_loss import *
+from encoder_decoder import *
+from optimizer_warmup import *
+from caip_dataset import *
 
 class ModelTrainer:
     """Wrapper Class around training model and data loader
